backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
google-generativeai==0.8.3
PyMuPDF==1.23.8
pillow==10.1.0
python-dotenv==1.0.0
werkzeug==3.0.1
python-multipart==0.0.6
gunicorn==21.2.0
psycopg2-binary==2.9.9
openpyxl==3.1.2